move sanitizeLabel to Util func and update imports
